refactor(cli): extract InlineDiff; fix FLEURS samples cap
- Move the duplicated `generateInlineDiff` edit-distance diff renderer
  (byte-identical copies in AsrBenchmark.swift and FleursBenchmark.swift,
  ~110 lines each) to a shared `InlineDiff.generate(reference:hypothesis:)`
  utility in `FluidAudioCLI/Utils/InlineDiff.swift`.
- Fix Devin Review finding in CohereBenchmark.swift line 263: when
  `--max-files` was omitted for FLEURS auto-download,
  `samplesPerLanguage: maxFiles ?? 100` silently capped each language at
  100 samples. Switch to `Int.max`, which is FLEURSBenchmark's documented
  sentinel for "download all available".
