+++ Sources/FluidAudioCLI/Commands/ASR/Parakeet/SlidingWindow/AsrBenchmark.swift
@@ -532,7 +532,8 @@ extension ASRBenchmark {
         let hypWords = normalizedHypothesis.components(separatedBy: .whitespacesAndNewlines).filter { !$0.isEmpty }
 
         // Generate inline diff
-        let (referenceDiff, hypothesisDiff) = generateInlineDiff(reference: refWords, hypothesis: hypWords)
+        let (referenceDiff, hypothesisDiff) = InlineDiff.generate(
+            reference: refWords, hypothesis: hypWords)
 
         logger.info("Normalized Reference:\t\(referenceDiff)")
         logger.info("Normalized Hypothesis:\t\(hypothesisDiff)")
@@ -621,116 +622,6 @@ extension ASRBenchmark {
         }
 
         return differences.reversed()  // Reverse to get correct order
-    }
-
-    /// Generate inline diff with full lines and highlighted differences
-    private func generateInlineDiff(reference: [String], hypothesis: [String]) -> (String, String) {
-        let m = reference.count
-        let n = hypothesis.count
-
-        // Handle empty hypothesis or reference
-        if n == 0 {
-            let supportsColor = ProcessInfo.processInfo.environment["TERM"] != nil
-            let redColor = supportsColor ? "\u{001B}[31m" : "["
-            let resetColor = supportsColor ? "\u{001B}[0m" : "]"
-            let refString = reference.map { "\(redColor)\($0)\(resetColor)" }.joined(separator: " ")
-            let hypString = ""
-            return (refString, hypString)
-        }
-        if m == 0 {
-            let supportsColor = ProcessInfo.processInfo.environment["TERM"] != nil
-            let greenColor = supportsColor ? "\u{001B}[32m" : "["
-            let resetColor = supportsColor ? "\u{001B}[0m" : "]"
-            let refString = ""
-            let hypString = hypothesis.map { "\(greenColor)\($0)\(resetColor)" }.joined(separator: " ")
-            return (refString, hypString)
-        }
-
-        // Create DP table for edit distance with backtracking
-        var dp = Array(repeating: Array(repeating: 0, count: n + 1), count: m + 1)
-
-        // Initialize base cases
-        for i in 0...m { dp[i][0] = i }
-        for j in 0...n { dp[0][j] = j }
-
-        // Fill DP table
-        for i in 1...m {
-            for j in 1...n {
-                if reference[i - 1] == hypothesis[j - 1] {
-                    dp[i][j] = dp[i - 1][j - 1]
-                } else {
-                    dp[i][j] =
-                        1
-                        + min(
-                            dp[i - 1][j],  // deletion
-                            dp[i][j - 1],  // insertion
-                            dp[i - 1][j - 1]  // substitution
-                        )
-                }
-            }
-        }
-
-        // Check if terminal supports colors
-        let supportsColor = ProcessInfo.processInfo.environment["TERM"] != nil
-        let redColor = supportsColor ? "\u{001B}[31m" : "["
-        let greenColor = supportsColor ? "\u{001B}[32m" : "["
-        let resetColor = supportsColor ? "\u{001B}[0m" : "]"
-
-        // Backtrack to identify differences
-        var i = m
-        var j = n
-        var refDiffWords: [(String, Bool)] = []  // (word, isDifferent)
-        var hypDiffWords: [(String, Bool)] = []  // (word, isDifferent)
-
-        while i > 0 || j > 0 {
-            if i > 0 && j > 0 && reference[i - 1] == hypothesis[j - 1] {
-                // Match
-                refDiffWords.insert((reference[i - 1], false), at: 0)
-                hypDiffWords.insert((hypothesis[j - 1], false), at: 0)
-                i -= 1
-                j -= 1
-            } else if i > 0 && j > 0 && dp[i][j] == dp[i - 1][j - 1] + 1 {
-                // Substitution
-                refDiffWords.insert((reference[i - 1], true), at: 0)
-                hypDiffWords.insert((hypothesis[j - 1], true), at: 0)
-                i -= 1
-                j -= 1
-            } else if i > 0 && dp[i][j] == dp[i - 1][j] + 1 {
-                // Deletion (word in reference but not in hypothesis)
-                refDiffWords.insert((reference[i - 1], true), at: 0)
-                i -= 1
-            } else if j > 0 && dp[i][j] == dp[i][j - 1] + 1 {
-                // Insertion (word in hypothesis but not in reference)
-                hypDiffWords.insert((hypothesis[j - 1], true), at: 0)
-                j -= 1
-            } else {
-                break
-            }
-        }
-
-        // Build the formatted strings
-        var refString = ""
-        var hypString = ""
-
-        for (word, isDifferent) in refDiffWords {
-            if !refString.isEmpty { refString += " " }
-            if isDifferent {
-                refString += "\(redColor)\(word)\(resetColor)"
-            } else {
-                refString += word
-            }
-        }
-
-        for (word, isDifferent) in hypDiffWords {
-            if !hypString.isEmpty { hypString += " " }
-            if isDifferent {
-                hypString += "\(greenColor)\(word)\(resetColor)"
-            } else {
-                hypString += word
-            }
-        }
-
-        return (refString, hypString)
     }
 }
